internal: minor include fixes and missed format
